+++ CHANGELOG.md
@@ -7,11 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Upgrade to Tailwind v2
+
 ## [1.0.1] - 2020-11-21
 
 ### Added
 
-- Add Windows code signing certificate
+- Windows code signing certificate
+
+### Changed
+
 - Update macOS icon for macOS Big Sur
 
 ## [1.0.0] - 2020-11-07