+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix bug where description was not cleared after time entries was stopped
+
 ## [1.1.0] - 2021-01-07
 
 ### Added