+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.3.1] - 2021-11-18
+
+- Update CSC
+
 ## [1.3.0] - 2021-08-17
 
 ### Changed